requirements.txt
Flask==3.0.0
pandas==2.1.0
numpy==1.26.4
scikit-learn==1.3.2
xgboost==2.0.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.5.2
Werkzeug==3.0.1
google-generativeai
python-dotenv
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
